Reenable package environments
Hopefully fixes #4010.

Parallel to this commit there will be a differential for a new command
line flag in GHC: `-no-implicit-package-env`. This flag explicitly
disables the implicit search for package.env files in GHC.

Before this patch GHC abused the `-hide-all-packages` flag to disable
package.env search which lead to the behaviour in #4010.